+++ src/manager/schemas/marketing/contact/delete.js
@@ -1,9 +1,6 @@
 /**
  * Schema for DELETE /marketing/contact
  */
-const { DEFAULT_PROVIDERS } = require('../../../libraries/email/constants.js');
-
 module.exports = () => ({
   email: { types: ['string'], default: undefined, required: true },
-  providers: { types: ['array'], default: DEFAULT_PROVIDERS },
 });